pyproject.toml
[tool.poetry]
name = "screenshare-ro-text"
version = "0.1.0"
description = ""
authors = []

packages = [
    { include = "screenshare_to_text" },
]

[tool.poetry.dependencies]
python = ">=3.10"
litellm = "^1.74.7"
opencv-python = "^4.12.0.88"
yarl = "^1.20.1"
python-dotenv = "^1.1.1"
numpy = "^2.2.6"
certifi = "^2025.7.14"
httpx = "^0.28.1"
sniffio = "^1.3.1"
idna = "^3.10"
httpcore = "^1.0.9"
h11 = "^0.16.0"
aiohttp = "^3.12.14"
multidict = "^6.6.3"
attrs = "^25.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
